pyproject.toml
[tool.poetry]
name = "constrained-attacks"
version = "0.1.2"
description = "constrained-attacks is a framework to generate adversarial examples under domain specific constraints."
authors = ["Thibault Simonetto <thibault.simonetto.001@student.uni.lu>"]
license = "MIT"
readme="README.md"

[tool.poetry.dependencies]
python = "~3.8"
numpy = "^1.22.3"
joblib = "^1.1.0"
pymoo = "^0.5.0"
tqdm = "^4.63.1"
pandas = "^1.4.1"
adversarial-robustness-toolbox = {version = "1.10", extras = ["tensorflow"], optional=true}
tensorflow = {version ="2.8", optional=true}
tensorflow-io-gcs-filesystem = {version ="0.25.0", optional=true}
protobuf = {version ="^3.20.0", optional=true}
torch = {version ="1.12.0", optional=true}
torchvision = {version ="0.13.0", optional=true}
serval-ml-commons = {path = "../mlc/", develop = true, extras = ["tabsurvey"] }
torchattacks = "^3.4.0"
seaborn = "^0.12.2"
comet-ml = "^3.33.8"

[tool.poetry.extras]
tensorflow = ["tensorflow", "adversarial-robustness-toolbox", "tensorflow-io-gcs-filesystem", "protobuf"]
pytorch = ["torch", "torchvision"]

[tool.poetry.group.dev.dependencies]
pytest = "7.1.2"
pytest-cov = "^3.0.0"
pre-commit = "^2.17.0"
flake8 = "^4.0.1"
black = {version = "^22.1.0", allow-prereleases = true}
mypy = "^0.941"
isort = "^5.10.1"
scikit-learn = "^1.0.2"
pip-licenses = "^3.5.3"

[tool.poetry.group.docs.dependencies]
sphinx-rtd-theme = "^1.2.1"
numpydoc = "^1.5.0"
sphinx = "^6.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
line_length = 79
